requirements.txt
httpx==0.27.0
scipy==1.11.2
pandas==1.5.3
pyautogen==0.2.33
portpy==1.0.4.1
cvxpy==1.3.2
numpy==1.26.0
matplotlib==3.8.0
seaborn==0.13.2
scikit-image==0.21.0
termcolor==2.4.0
joblib==1.4.2
xgboost==2.0.3
google-generativeai==0.7.2
